backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
gunicorn==25.0.3
psycopg[binary]==3.3.2
